refactor: move card view feature (#876)
