+++ src/pages/card-list/ui/CardListPage.tsx
@@ -7,7 +7,7 @@ import { type Deck, useDeck } from "@/entities/deck";
 import { type Preferences, usePreferences } from "@/entities/preferences";
 import { useCardReadState } from "@/features/card/read";
 import { CardList } from "@/features/card-list";
-import { BackText } from "@/features/card/view";
+import { BackText } from "@/features/card-view";
 import { DeckStartForm, useDeckFilterState, useEditStudyProgress, useStudyCards } from "@/features/study";
 import { RemoteReadBoundary } from "@/shared/ui/remote-read-boundary";
 import { RouteFeedback } from "@/shared/ui/route-feedback";
